+++ src/core/moves/admixture/AdmixtureEdgeReplaceResidualsFNPR.cpp
@@ -133,8 +133,8 @@ double AdmixtureEdgeReplaceResidualsFNPR::performSimpleMove( void ) {
     RandomNumberGenerator* rng     = GLOBAL_RNG;
     
     // stage FNPR pointers
-    int numIntNodes = tau.getNumberOfInteriorNodes();
-    int numTaxa = tau.getNumberOfTips();
+    size_t numIntNodes = tau.getNumberOfInteriorNodes();
+    size_t numTaxa = tau.getNumberOfTips();
     pruneNode = NULL;
     while (pruneNode == NULL || pruneNode == &tau.getRoot() || pruneNode->getNumberOfChildren() != 2)
     {
